pyproject.toml
[project]
name = "cripto-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cachetools>=5.5.2",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "pytelegrambotapi>=4.28.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "yfinance>=0.2.65",
]
